fix: improve handling of overflows when there are more pack than we can hold.
Internally, there is a statically allocated vec which holds opened packs and
indices.
When reconciling the disk-state with what's currently loaded, it was possible
to get into a situation where there were more files than we could fit into
the slotmap.

The system now fails loudly with a specific Error so callers can deal with it.